Fix error

+++ reunhangout/plenaries/consumers.py
@@ -47,10 +47,11 @@ class PlenaryConsumer(WebsocketConsumer):
             return self.handle_error('Plenary not found')
 
         # Remove previous presence if the user is reconnecting
-        Presence.objects.filter(
-            room__channel_name=plenary.channel_group_name,
-            user=self.scope['user']
-        ).delete()
+        if self.scope['user'].is_authenticated:
+            Presence.objects.filter(
+                room__channel_name=plenary.channel_group_name,
+                user=self.scope['user']
+            ).delete()
 
         # Handle max connections
         if plenary.max_participants > 0 and not plenary.has_admin(self.scope['user']):
